Defining internal package for all the modules and moving all packages except auxiliary to internal.
