+++ modules/assets/internal/genesis/codec.go
@@ -7,7 +7,7 @@ package genesis
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/persistenceOne/persistenceSDK/modules/splits/mapper"
+	"github.com/persistenceOne/persistenceSDK/modules/assets/internal/mapper"
 	"github.com/persistenceOne/persistenceSDK/schema"
 )
