+++ modules/assets/internal/transactions/burn/constants.go
@@ -3,11 +3,11 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-package mint
+package burn
 
-import "github.com/persistenceOne/persistenceSDK/modules/assets/mapper"
+import "github.com/persistenceOne/persistenceSDK/modules/assets/internal/mapper"
 
-const TransactionName = "mint"
+const TransactionName = "burn"
 const TransactionRoute = mapper.TransactionRoute + "/" + TransactionName
 const TransactionShort = ""
 const TransactionLong = ""
