+++ modules/assets/internal/transactions/define/transaction.go
@@ -7,7 +7,7 @@ package define
 
 import (
 	"github.com/persistenceOne/persistenceSDK/constants/flags"
-	"github.com/persistenceOne/persistenceSDK/modules/orders/mapper"
+	"github.com/persistenceOne/persistenceSDK/modules/assets/internal/mapper"
 	"github.com/persistenceOne/persistenceSDK/schema/helpers"
 	"github.com/persistenceOne/persistenceSDK/schema/helpers/base"
 )
