+++ modules/assets/internal/transactions/mint/constants.go
@@ -3,11 +3,11 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-package mutate
+package mint
 
-import "github.com/persistenceOne/persistenceSDK/modules/assets/mapper"
+import "github.com/persistenceOne/persistenceSDK/modules/assets/internal/mapper"
 
-const TransactionName = "mutate"
+const TransactionName = "mint"
 const TransactionRoute = mapper.TransactionRoute + "/" + TransactionName
 const TransactionShort = ""
 const TransactionLong = ""
