+++ modules/assets/internal/transactions/mutate/constants.go
@@ -3,11 +3,11 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-package define
+package mutate
 
-import "github.com/persistenceOne/persistenceSDK/modules/assets/mapper"
+import "github.com/persistenceOne/persistenceSDK/modules/assets/internal/mapper"
 
-const TransactionName = "define"
+const TransactionName = "mutate"
 const TransactionRoute = mapper.TransactionRoute + "/" + TransactionName
 const TransactionShort = ""
 const TransactionLong = ""
