+++ modules/maintainers/internal/genesis/codec.go
@@ -7,7 +7,7 @@ package genesis
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/persistenceOne/persistenceSDK/modules/metas/mapper"
+	"github.com/persistenceOne/persistenceSDK/modules/maintainers/internal/mapper"
 	"github.com/persistenceOne/persistenceSDK/schema"
 )
