+++ modules/maintainers/internal/queries/maintainer/codec.go
@@ -3,22 +3,23 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-package genesis
+package maintainer
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/persistenceOne/persistenceSDK/modules/orders/mapper"
+	"github.com/persistenceOne/persistenceSDK/modules/maintainers/internal/mapper"
 	"github.com/persistenceOne/persistenceSDK/schema"
 )
 
-func (genesisState) RegisterCodec(codec *codec.Codec) {
-	codec.RegisterConcrete(genesisState{}, mapper.ModuleRoute+"/"+"genesisState", nil)
+func registerCodec(codec *codec.Codec) {
+	codec.RegisterConcrete(queryRequest{}, QueryRoute+"/"+"request", nil)
+	codec.RegisterConcrete(queryResponse{}, QueryRoute+"/"+"response", nil)
 }
 
 var packageCodec = codec.New()
 
 func init() {
-	State.RegisterCodec(packageCodec)
+	registerCodec(packageCodec)
 	schema.RegisterCodec(packageCodec)
 	mapper.Mapper.RegisterCodec(packageCodec)
 	packageCodec.Seal()
