+++ modules/orders/internal/genesis/codec.go
@@ -7,7 +7,7 @@ package genesis
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/persistenceOne/persistenceSDK/modules/maintainers/mapper"
+	"github.com/persistenceOne/persistenceSDK/modules/orders/internal/mapper"
 	"github.com/persistenceOne/persistenceSDK/schema"
 )
