+++ modules/splits/internal/genesis/codec.go
@@ -1,0 +1,25 @@
+/*
+ Copyright [2019] - [2020], PERSISTENCE TECHNOLOGIES PTE. LTD. and the persistenceSDK contributors
+ SPDX-License-Identifier: Apache-2.0
+*/
+
+package genesis
+
+import (
+	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/persistenceOne/persistenceSDK/modules/splits/internal/mapper"
+	"github.com/persistenceOne/persistenceSDK/schema"
+)
+
+func (genesisState) RegisterCodec(codec *codec.Codec) {
+	codec.RegisterConcrete(genesisState{}, mapper.ModuleRoute+"/"+"genesisState", nil)
+}
+
+var packageCodec = codec.New()
+
+func init() {
+	State.RegisterCodec(packageCodec)
+	schema.RegisterCodec(packageCodec)
+	mapper.Mapper.RegisterCodec(packageCodec)
+	packageCodec.Seal()
+}
