+++ modules/splits/internal/queries/split/codec.go
@@ -3,12 +3,11 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-package maintainer
+package split
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
-	assetsMapper "github.com/persistenceOne/persistenceSDK/modules/assets/mapper"
-	"github.com/persistenceOne/persistenceSDK/modules/maintainers/mapper"
+	"github.com/persistenceOne/persistenceSDK/modules/splits/internal/mapper"
 	"github.com/persistenceOne/persistenceSDK/schema"
 )
 
@@ -22,7 +21,6 @@ var packageCodec = codec.New()
 func init() {
 	registerCodec(packageCodec)
 	schema.RegisterCodec(packageCodec)
-	assetsMapper.Mapper.RegisterCodec(packageCodec)
 	mapper.Mapper.RegisterCodec(packageCodec)
 	packageCodec.Seal()
 }
